Extract and fix register receive nft tests
