+++ src/lib.rs
@@ -21,6 +21,7 @@ mod unittest_public_metadata;
 mod unittest_queries;
 mod unittest_reveal;
 mod unittest_transfer;
+mod unittest_register_receive_nft;
 mod unittest_send;
 mod unittest_whitelisted_approval;
 mod utils;
